added explosion particles, added chef upgrades + balancing to upgrade stats/updrage and sell costs

+++ Assets/Scripts/Enemy/Enemy.cs
@@ -180,7 +180,7 @@ public class Enemy : MonoBehaviour
     }
 
     // Reduce hunger when given food
-    public void ApplyFood(int damage)
+    public void ApplyFood(float damage)
     {
         CurrentHunger -= damage;
     }
